+++ index.js
@@ -1,30 +1,27 @@
 module.exports = function (broccoli) {
-  CoffeeScriptPreprocessor.prototype = Object.create(broccoli.Preprocessor.prototype)
-  CoffeeScriptPreprocessor.prototype.constructor = CoffeeScriptPreprocessor
-  function CoffeeScriptPreprocessor (options) {
+  CoffeeScriptFilter.prototype = Object.create(broccoli.Filter.prototype)
+  CoffeeScriptFilter.prototype.constructor = CoffeeScriptFilter
+  function CoffeeScriptFilter (options) {
     this.options = options || {}
   }
 
-  CoffeeScriptPreprocessor.prototype.extensions = ['coffee']
-  CoffeeScriptPreprocessor.prototype.targetExtension = 'js'
+  CoffeeScriptFilter.prototype.extensions = ['coffee']
+  CoffeeScriptFilter.prototype.targetExtension = 'js'
 
-  CoffeeScriptPreprocessor.prototype.processString = function (string, info, callback) {
+  CoffeeScriptFilter.prototype.processString = function (string, info) {
     // We must be careful to create a fresh options hash every time.
     // https://github.com/jashkenas/coffee-script/issues/1924#issuecomment-28157026
     var options = {
       bare: this.options.bare
     }
-    var output
     try {
-      output = require('coffee-script').compile(string, options)
+      return require('coffee-script').compile(string, options)
     } catch (err) {
       err.line = err.location && err.location.first_line
       err.column = err.location && err.location.first_column
-      callback(err)
-      return
+      throw err
     }
-    callback(null, output)
   }
 
-  return CoffeeScriptPreprocessor
+  return CoffeeScriptFilter
 }