+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Use new `Filter` architecture
+
 # 0.0.1
 
 * Initial release